requirements.txt
astroid==2.3.3
autopep8==1.5.2
click==7.1.1
Flask==1.1.2
gitdb==4.0.4
GitPython==3.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
ply==3.11
pycodestyle==2.5.0
pylint==2.4.4
six==1.14.0
smmap==3.0.2
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
